requirements.txt
asgiref==3.7.2
atpublic==4.0
bangla==0.0.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
Django==5.0.2
django-browser-reload==1.12.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-rest-framework==0.1.0
django-tailwind==3.8.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
env==0.1.0
environ==1.0
environment==1.0.0
filelock==3.13.1
fonttools==4.53.0
gunicorn==21.2.0
idna==3.6
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mysqlclient==2.2.4
numpy==1.26.4
packaging==24.0
pillow==10.2.0
platformdirs==4.2.0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.1
world==5.1.1
